Make bio work with prefilled input text
